Tratamento de exceção - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exeptions.DatabaseExeption;
 import com.educandoweb.course.services.exeptions.ResourceNotFoundExeption;
 
 @Service
@@ -30,7 +33,15 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		if(!userRepository.existsById(id)) {
+			throw new ResourceNotFoundExeption(id);
+		}
+		try {
+			userRepository.deleteById(id);
+		}  catch (DataIntegrityViolationException e) {
+			throw new DatabaseExeption(e.getMessage());
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
